Fix tests after refactoring
