+++ src/helpers/phoneHelper.js
@@ -1,4 +1,5 @@
 module.exports.formatMobile = function (phoneNumber) {
-  let numberMatch = /^\+?(44)?0?(\d+)$/;
-  return phoneNumber.replace(numberMatch, '0$2');
+  let spaceMatcher = /\s/g;
+  let numberMatcher = /^\+?(44)?0?(\d+)$/;
+  return phoneNumber.replace(spaceMatcher, '').replace(numberMatcher, '0$2');
 }
