+++ test/suite/_setup.test.js
@@ -1,4 +1,4 @@
-const yotiService = require('../services/yotiService.js');
+const yotiService = require('../../services/yotiService.js');
 const mockYotiClient = require('../helpers/mockYotiClient.js');
 
 yotiService.getClient = function() {
